--- Final/templates/ERA.html
@@ -1,129 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>  
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>MLB ERA v Wins</title>
-    
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/d3/5.5.0/d3.min.js"></script>
-    <script src="https://cdn.plot.ly/plotly-latest.min.js"></script>
-    
-    <link rel="stylesheet" type="text/css" href="stylesheet.css" media="screen" >
-</head>
-
-
-
-<body style="background-color:teal">
-    
-    <div class="W3-container" height="img-fluid" style="background-color:black">
-                   
-        <!-- Row 1 -->
-         <div class="row">
-            <div class="col-xs-2 col-sm-3 col-md-3" class="img-fluid" height="auto"></div>
-            <div class="col-xs-2 col-sm-3 col-md-3" class="img-fluid" height="auto">
-                 <h1> <a href="/"><span style="color:white">MLB</span></a> </h1>
-            </div>
-            
-            
-            <div class="col-xs-3 col-sm-3 col-md-3" class="img-fluid">
-            </div>                        
-            <div class="col-xs-2 col-sm-2 col-md-2" class="img-fluid">   
-                <div class="dropdown">
-                    <button class="btn btn-secondary dropdown-toggle" type="button" id="dropdownMenuButton" data-toggle="dropdown" >
-                        Stats
-                                </button>
-                                <div class="dropdown-menu" aria-labelledby="dropdownMenuButton">
-                                    <a class="dropdown-item" href="runs">Runs</a>
-                                    <a class="dropdown-item" href="homeruns">Homeruns</a>
-                                    <a class="dropdown-item" href="ERA">ERA</a>
-                                </div>
-                            </div>
-                        </div> 
-        </div>
-        
-    </div>
-        <hr>        
-    <div class="container">
-        <div class="class jumbotron">
-            
-            <!-- Row 1 Summary-->
-            <div class="row">
-                
-                <div class="col-md-12">
-                    <h2>ERA v. Wins</h2>
-                    <hr>
-                </div>
-                    
-                <div class="col-md-4">
-                    <P>
-                        <img src="{{url_for('static', filename='images.jpg')}}" class="img-fluid"/>
-                </div>
-                <div class="col-md-8">                            
-            
-                            In baseball statistics, earned run average (ERA) is the mean of earned
-                            runs given up by a pitcher per nine innings pitched (i.e. the traditional 
-                            length of a game). It is determined by dividing the number of earned runs 
-                            allowed by the number of innings pitched and multiplying by nine. Runs 
-                            resulting from defensive errors (including pitchers' defensive errors) 
-                            are recorded as unearned runs and omitted from ERA calculations.
-                </div>
-                <div class="col-md-12">           
-                        <br>
-                        <br>                 
-                    <p>Enter a team ERA total in decimal form to the hundredths.  For context the lowest 
-                        team ERA in the last 50 years is 2.53 in a season.</p>
-                    </div>
-                
-            
-
-            <div class="container">
-                <div class="col-md-6"></div>
-                <div class="col-md-6">
-                  <div class="thumbnail">
-                    <div class="">
-                        
-                        <form method="POST">
-                        
-                        <input type="number" step="0.01" name="ERA">
-                        
-                        <button type="submit" class="btn btn-primary">Prediction</button>
-
-                      </form>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            <br>
-            <br>
-            <hr>
-            <div class="container">
-                
-            <h3>Based on a team ERA of {{ input }} in a season the model predicts that they will win {{ wins }} games.</h3>
-            </div>
-            
-            <div class="container">
-                <div class="col-md-12">
-                  <div id="ERAScatter"></div>
-                </div>
-              </div> 
-            </section>   
-            </div>
-            
-        </div>
-
-              
-    </div>
-
-       
-    
-    <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49" crossorigin="anonymous"></script>
-    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js" integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy" crossorigin="anonymous"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/d3/5.5.0/d3.js"></script>
-    <script src="https://cdn.plot.ly/plotly-latest.min.js"></script>
-    <script src="{{ url_for('static', filename='js/plots.js') }}"></script>
-</body>
-
-</html>
