+++ templates/stylesheet.css
@@ -1,0 +1,22 @@
+    h1 {
+        color: white;
+        font-size: 2em;
+        background-color: black;
+        text-align: center;
+        }   
+
+    h2 {
+        color:teal;
+        font-size: 2em;
+        background-color:white;
+        text-align: center;
+        }
+    h3 {
+        color:teal;
+        font-size: 2em;
+        background-color:white;
+        text-align: center;
+        }
+    .jumbotron {
+            background:white
+        }
